refactor: constant.js 디렉터리 변경(utils -> constant)
